+++ app/src/main/java/ru/touchthegrass/tpc/data/local/LocalProvider.kt
@@ -1,0 +1,7 @@
+package ru.touchthegrass.tpc.data.local
+
+abstract class LocalProvider {
+
+    private var id = 1
+    protected fun getId() = id++
+}
