+++ app/src/main/java/ru/touchthegrass/tpc/manager/local/LocalLobbyManager.kt
@@ -1,21 +1,26 @@
 package ru.touchthegrass.tpc.manager.local
 
+import kotlinx.coroutines.CoroutineScope
 import ru.touchthegrass.tpc.data.Lobby
 import ru.touchthegrass.tpc.data.PieceColor
 import ru.touchthegrass.tpc.data.PlayerGameSession
 import ru.touchthegrass.tpc.data.PlayerStatus
-import ru.touchthegrass.tpc.data.local.LocalLobbyProvider
+import ru.touchthegrass.tpc.data.local.LocalGameSessionProvider
 import ru.touchthegrass.tpc.data.local.LocalPlayerProvider
 import ru.touchthegrass.tpc.manager.LobbyManager
+import ru.touchthegrass.tpc.repository.LobbyRepository
 
-class LocalLobbyManager : LobbyManager {
+class LocalLobbyManager(
+    val coroutineScope: CoroutineScope,
+    override val lobbyRepository: LobbyRepository
+) : LobbyManager {
 
     override fun createLobby(): Lobby {
-        return LocalLobbyProvider.createLobby()
+        return lobbyRepository.createLobby(Lobby(gameSession = LocalGameSessionProvider.createGameSession()))
     }
 
     override fun connectPlayer(gameSessionId: Int, playerId: Int): Lobby {
-        val lobby = LocalLobbyProvider.getLobbyByGameSessionId(gameSessionId)
+        val lobby = lobbyRepository.getLobbyByGameSessionId(gameSessionId)
         val player = LocalPlayerProvider.getPlayerById(playerId)
         lobby.playerInfos.add(
             PlayerGameSession(
@@ -27,12 +32,12 @@ class LocalLobbyManager : LobbyManager {
     }
 
     override fun disconnectPlayer(gameSessionId: Int, playerId: Int) {
-        val lobby = LocalLobbyProvider.getLobbyByGameSessionId(gameSessionId)
+        val lobby = lobbyRepository.getLobbyByGameSessionId(gameSessionId)
         lobby.playerInfos.removeIf { it.player.id == playerId }
     }
 
     override fun setPlayerColor(gameSessionId: Int, playerId: Int, pieceColor: PieceColor): List<PlayerGameSession> {
-        val lobby = LocalLobbyProvider.getLobbyByGameSessionId(gameSessionId)
+        val lobby = lobbyRepository.getLobbyByGameSessionId(gameSessionId)
         val updatedInfos = lobby.playerInfos
             .map { if (it.player.id == playerId) it.copy(color = pieceColor) else it }
             .toMutableList()
@@ -40,8 +45,12 @@ class LocalLobbyManager : LobbyManager {
         return updatedInfos
     }
 
-    override fun setPlayerStatus(gameSessionId: Int, playerId: Int, status: PlayerStatus): List<PlayerGameSession> {
-        val lobby = LocalLobbyProvider.getLobbyByGameSessionId(gameSessionId)
+    override fun setPlayerStatus(
+        gameSessionId: Int,
+        playerId: Int,
+        status: PlayerStatus
+    ): List<PlayerGameSession> {
+        val lobby = lobbyRepository.getLobbyByGameSessionId(gameSessionId)
         val updatedInfos = lobby.playerInfos
             .map { if (it.player.id == playerId) it.copy(status = status) else it }
             .toMutableList()
