--- app/src/main/java/ru/touchthegrass/tpc/ui/component/GameRuleDrawerContent.kt
@@ -1,73 +1,0 @@
-package ru.touchthegrass.tpc.ui.component
-
-import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.verticalScroll
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.MenuOpen
-import androidx.compose.material3.*
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.layout.Layout
-import androidx.compose.ui.layout.layoutId
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.unit.dp
-import ru.touchthegrass.tpc.R
-import ru.touchthegrass.tpc.data.GameRule
-import ru.touchthegrass.tpc.ui.util.LayoutType
-import ru.touchthegrass.tpc.ui.util.tpcDrawerMeasurePolicy
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun LobbyDrawerContent(
-    rules: List<GameRule>,
-    closeDrawer: () -> Unit
-) {
-    ModalDrawerSheet {
-        Layout(
-            modifier = Modifier.padding(16.dp),
-            content = {
-                Column(
-                    modifier = Modifier.layoutId(LayoutType.HEADER),
-                    horizontalAlignment = Alignment.CenterHorizontally,
-                    verticalArrangement = Arrangement.spacedBy(8.dp)
-                ) {
-                    Row(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(16.dp),
-                        horizontalArrangement = Arrangement.SpaceBetween,
-                        verticalAlignment = Alignment.CenterVertically
-                    ) {
-                        Text(
-                            text = stringResource(id = R.string.game_rules).uppercase(),
-                            style = MaterialTheme.typography.titleMedium,
-                            color = MaterialTheme.colorScheme.primary
-                        )
-                        IconButton(onClick = closeDrawer) {
-                            Icon(
-                                imageVector = Icons.Default.MenuOpen,
-                                contentDescription = stringResource(id = R.string.lobby_drawer)
-                            )
-                        }
-                    }
-                }
-                Column(
-                    modifier = Modifier
-                        .layoutId(LayoutType.CONTENT)
-                        .verticalScroll(rememberScrollState()),
-                    horizontalAlignment = Alignment.Start
-                ) {
-                    for (rule in rules) {
-                        GameRuleChip(
-                            icon = rule.icon,
-                            title = rule.value
-                        )
-                    }
-                }
-            },
-            measurePolicy = tpcDrawerMeasurePolicy
-        )
-    }
-}
